+++ src/successModal/SuccessModal.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
-import { shallow, mount } from 'enzyme'
+import { mount, shallow } from 'enzyme'
 
-import SuccessModal from 'successModal'
+import SuccessModal from './index'
 
 const defaultProps = {
   isOpen: false,
@@ -9,53 +9,49 @@ const defaultProps = {
   imageSrc: 'https://svgshare.com/i/AGz.svg',
   imageText: 'Illustration description',
   closeOnEsc: true,
+  onClose: () => {},
+  forwardedRef: null,
 }
 
 describe('<SuccessModal>', () => {
   let mockClose
-  let wrapper
   let wrapperOpen
 
-  beforeEach(() => {
+  it('Should be visible if isOpen is set to true', () => {
     mockClose = jest.fn()
-    wrapper = shallow(<SuccessModal {...defaultProps} />)
     wrapperOpen = mount(
       <SuccessModal {...defaultProps} isOpen onClose={mockClose}>
         Success description
       </SuccessModal>,
     )
-  })
-
-  it('Should be not visible if isOpen is set to false', () => {
-    expect(wrapper.find('.kirk-modal-dialog').exists()).toBe(false)
-  })
-
-  it('Should be visible if isOpen is set to true', () => {
     expect(wrapperOpen.find('.kirk-modal-dialog').exists()).toBe(true)
   })
 
+  it('Should be hidden if isOpen is set to false', () => {
+    const wrapper = shallow(<SuccessModal {...defaultProps} />)
+    expect(wrapper.find('.kirk-modal-dialog').exists()).toBe(false)
+  })
+
   it('Should have proper linked id to the content text', () => {
-    const ariaLabelledByValue = wrapperOpen.find('.kirk-successModal').prop('aria-labelledby')
-    expect(wrapperOpen.find(`#${ariaLabelledByValue} h1`).text()).toEqual('Success description')
+    expect(
+      wrapperOpen
+        .find('[data-test="success-title"]')
+        .at(0)
+        .text(),
+    ).toEqual('Success description')
   })
 
   it('Should have a confirmation button and call the according function when click on it', () => {
-    expect(wrapperOpen.find('.kirk-button-secondary').text()).toBe('Confirm')
-  })
-
-  it('Should have the proper image', () => {
-    const image = wrapperOpen.find('.kirk-successModal-image')
-    expect(image.prop('src')).toBe('https://svgshare.com/i/AGz.svg')
-    expect(image.prop('alt')).toBe('Illustration description')
-  })
-
-  it("Shouldn't access the image with screen readers", () => {
-    const image = wrapperOpen.find('.kirk-successModal-image')
-    expect(image.prop('aria-hidden')).toBe(true)
+    expect(
+      wrapperOpen
+        .find('[data-test="success-button"]')
+        .at(0)
+        .text(),
+    ).toBe('Confirm')
   })
 
   it('Should have a confirmation button and call the according function when click on it', () => {
-    const confirmButton = wrapperOpen.find('.kirk-button-secondary')
+    const confirmButton = wrapperOpen.find('[data-test="success-button"]').at(0)
     expect(confirmButton.text()).toBe('Confirm')
     confirmButton.simulate('click')
     expect(mockClose).toHaveBeenCalledTimes(1)